Prod-0: Move security config from main class to its own to help with testing
