+++ src/main/java/com/indiraactive/fulfillmentplatform/config/SecurityConfig.java
@@ -1,0 +1,17 @@
+package com.indiraactive.fulfillmentplatform.config;
+
+import org.springframework.boot.autoconfigure.security.oauth2.client.EnableOAuth2Sso;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@Configuration
+@EnableOAuth2Sso
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http.authorizeRequests()
+                .antMatchers("/health").permitAll()
+                .anyRequest().authenticated();
+    }
+}
